+++ app/javascript/packs/application.js
@@ -6,5 +6,6 @@ import { initMapbox } from '../plugins/init_mapbox';
 
 import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
 
+import { initAutocomplete } from '../plugins/init_autocomplete';
 
 initMapbox();
