+++ app/javascript/plugins/init_autocomplete.js
