+++ app/javascript/plugins/init_mapbox.js
@@ -1,0 +1,50 @@
+import mapboxgl from 'mapbox-gl';
+import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
+
+
+const fitMapToMarkers = (map, markers) => {
+  const bounds = new mapboxgl.LngLatBounds();
+  markers.forEach(marker => bounds.extend([ marker.lng, marker.lat ]));
+  map.fitBounds(bounds, { padding: 70, maxZoom: 15, duration: 0 });
+};
+
+const initMapbox = () => {
+  const mapElement = document.getElementById('map');
+    if (mapElement) { // only build a map if there's a div#map to inject into
+      mapboxgl.accessToken = mapElement.dataset.mapboxApiKey;
+      const map = new mapboxgl.Map({
+        container: 'map',
+        style: 'mapbox://styles/mapbox/streets-v10',
+  });
+    const markers = JSON.parse(mapElement.dataset.markers);
+    const addMarkersToMap = (map, markers) => {
+        markers.forEach((marker) => {
+          const popup = new mapboxgl.Popup().setHTML(marker.infoWindow);
+
+           const element = document.createElement('div');
+            element.className = 'marker';
+            element.style.backgroundImage = `url('${marker.image_url}')`;
+            element.style.backgroundSize = 'contain';
+            element.style.width = '25px';
+            element.style.height = '25px';
+
+          new mapboxgl.Marker(element)
+            .setLngLat([ marker.lng, marker.lat ])
+            .setPopup(popup) // add this
+            .addTo(map);
+
+    });
+};
+addMarkersToMap(map,markers);
+fitMapToMarkers(map,markers);
+map.addControl(new MapboxGeocoder({ accessToken: mapboxgl.accessToken,
+                                      mapboxgl: mapboxgl }));
+
+
+
+  }
+};
+
+
+export { initMapbox };
+
