+++ app/models/location.rb
@@ -4,4 +4,6 @@ class Location < ApplicationRecord
 
   validates :name, presence: true
   validates :address, presence: true
+
+
 end
