+++ app/views/gamenights/_info_window.html.erb
@@ -1,2 +1,2 @@
-<h2><%= gamenight.name %></h2>
-<p><%= gamenight.address %></p>
+<p><%= gamenight.boardgame.name %></p>
+<p><%= gamenight.location.address %></p>
